requirements.txt
pandas
requests
tqdm
numpy
scikit-learn
tensorflow
google-generativeai
python-dotenv
google-adk